Major changes in layout
The resistor arrays proved hard to solder, so I increased
the pad size. Due to the tight space constraints, this led
to relatively big changes in the board. One consequence is
that traces are now 8mil - a compromise between the 6mil
that seemed too fine and the 10mil that just wont fit.